_events_producers/requirements.txt
azure-eventhub
Faker